requirements.txt

pywa
sqlalchemy>=1.4
databases
asyncpg
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
argon2-cffi
python-jose[cryptography]
python-multipart
email-validator
httpx
aioredis
redis
apscheduler
typing-extensions
pytest
pytest-asyncio
pytest-cov
coverage
Cython<3.0
PyYAML==6.0.1